# DCS/implementation/backend/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: DCS API Server
    version: 0.0.1
servers:
    - url: http://0.0.0.0:8991
paths:
...
    /template/retrieve:
        get:
            tags:
                - TemplateRepository
            summary: retrieve TemplateRepository
            description: load submitted template and history/provenance summary. fetch reviewed template with metadata, review history, and validation results. fetch all template entries for dashboard view.
            operationId: TemplateRepository#retrieve
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ContractTemplateRetrieveResponse'
                                example:
                                    - created_at: Quasi id autem.
                                      created_by: Cumque et beatae assumenda architecto quaerat.
                                      description: Aspernatur occaecati omnis eveniet ea iure.
                                      did: Eos non temporibus non qui dolorem.
                                      document_number: 4166337592319625211
                                      meta_data: Repellat culpa ducimus et.
                                      name: Impedit impedit est voluptas voluptas id.
                                      state: Occaecati perspiciatis cumque.
                                      version: 9122291519645283215
                                    - created_at: Quasi id autem.
                                      created_by: Cumque et beatae assumenda architecto quaerat.
                                      description: Aspernatur occaecati omnis eveniet ea iure.
                                      did: Eos non temporibus non qui dolorem.
                                      document_number: 4166337592319625211
                                      meta_data: Repellat culpa ducimus et.
                                      name: Impedit impedit est voluptas voluptas id.
                                      state: Occaecati perspiciatis cumque.
                                      version: 9122291519645283215
                            example:
                                - created_at: Quasi id autem.
                                  created_by: Cumque et beatae assumenda architecto quaerat.
                                  description: Aspernatur occaecati omnis eveniet ea iure.
                                  did: Eos non temporibus non qui dolorem.
                                  document_number: 4166337592319625211
                                  meta_data: Repellat culpa ducimus et.
                                  name: Impedit impedit est voluptas voluptas id.
                                  state: Occaecati perspiciatis cumque.
                                  version: 9122291519645283215
                                - created_at: Quasi id autem.
                                  created_by: Cumque et beatae assumenda architecto quaerat.
                                  description: Aspernatur occaecati omnis eveniet ea iure.
                                  did: Eos non temporibus non qui dolorem.
                                  document_number: 4166337592319625211
                                  meta_data: Repellat culpa ducimus et.
                                  name: Impedit impedit est voluptas voluptas id.
                                  state: Occaecati perspiciatis cumque.
                                  version: 9122291519645283215
                "400":
                    description: 'bad_request: Bad request'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal_error: Internal server error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
    /template/retrieve/{template_id}:
        get:
            tags:
                - TemplateRepository
            summary: retrieve_by_id TemplateRepository
            description: Retrieve a template by template id.
            operationId: TemplateRepository#retrieve_by_id
            parameters:
                - name: template_id
                  in: path
                  description: ID of the contract template
                  required: true
                  schema:
                    type: string
                    description: ID of the contract template
                    example: Omnis suscipit.
                  example: In quis modi molestiae saepe.
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ContractTemplateRetrieveByIDResponse'
                            example:
                                created_at: Nihil porro qui totam.
                                created_by: Mollitia at.
                                description: Laudantium magni nobis perspiciatis nesciunt ut.
                                did: Dicta voluptatem est dolor ut.
                                document_number: 281758007857943350
                                meta_data: Voluptatem natus illo.
                                name: Earum illo dolor.
                                state: Aliquid deleniti rerum.
                                version: 1088163813797186908
                "400":
                    description: 'bad_request: Bad request'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
                "500":
                    description: 'internal_error: Internal server error'
                    content:
                        application/vnd.goa.error:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
    schemas:
...
        ContractTemplateRetrieveByIDResponse:
            type: object
            properties:
                created_at:
                    type: string
                    description: The timestamp when the contract template was created
                    example: Nihil doloribus nisi.
                created_by:
                    type: string
                    description: Identifier of who created the contract template
                    example: Eaque corporis aut consequatur unde voluptatem.
                description:
                    type: string
                    description: A description for that template
                    example: Velit odio.
                did:
                    type: string
                    description: Decentralized Identifier of the contract template
                    example: Consequatur quam facilis earum provident.
                document_number:
                    type: integer
                    description: The document number of the contract template
                    example: 9078570089367054262
                    format: int64
                meta_data:
                    description: The metadata of the contract template
                    example: Atque aspernatur neque et iusto quia.
                name:
                    type: string
                    description: The name of the contract template
                    example: Enim vero eos ut.
                state:
                    type: string
                    description: The state of the contract template
                    example: Quia non quia delectus corporis perspiciatis.
                version:
                    type: integer
                    description: The version number of the contract template
                    example: 1290277735002454176
                    format: int64
            description: Result for retrieving a contract template by id
            example:
                created_at: Dolore corrupti delectus asperiores autem aut iusto.
                created_by: Voluptatem velit sed vero occaecati.
                description: Vitae debitis.
                did: Mollitia eveniet et aliquam.
                document_number: 1056781419160284888
                meta_data: Eos animi laborum veritatis cupiditate.
                name: Sunt veritatis et et suscipit dolores.
                state: Totam similique dolorum reiciendis.
                version: 4650894262499133558
            required:
                - did
                - document_number
                - version
                - state
                - created_by
                - created_at
                - meta_data
        ContractTemplateRetrieveResponse:
            type: object
            properties:
                created_at:
                    type: string
                    description: The timestamp when the contract template was created
                    example: Necessitatibus vel.
                created_by:
                    type: string
                    description: Identifier of who created the contract template
                    example: Veritatis vel.
                description:
                    type: string
                    description: A description for that template
                    example: Laborum vitae et non enim quo.
                did:
                    type: string
                    description: Decentralized Identifier of the contract template
                    example: Consectetur fugit repellendus.
                document_number:
                    type: integer
                    description: The document number of the contract template
                    example: 841325136931723046
                    format: int64
                meta_data:
                    description: The metadata of the contract template
                    example: Commodi omnis ullam beatae beatae.
                name:
                    type: string
                    description: The name of the contract template
                    example: Consequatur adipisci autem ut facilis earum.
                state:
                    type: string
                    description: The state of the contract template
                    example: A alias.
                version:
                    type: integer
                    description: The version number of the contract template
                    example: 3673781195744267971
                    format: int64
            description: Result for retrieving a contract template
            example:
                created_at: Voluptatem alias dolores architecto provident tenetur.
                created_by: Ullam eos.
                description: Consequuntur nesciunt.
                did: Dolor autem sed architecto sint.
                document_number: 4868152360019307510
                meta_data: Ut autem aut.
                name: Amet dolor provident quia velit.
                state: Maiores laudantium cumque asperiores.
                version: 2603452809852102205
            required:
                - did
                - document_number
                - version
                - state
                - created_by
                - created_at
                - meta_data
...
        Error:
            type: object
            properties:
                fault:
                    type: boolean
                    description: Is the error a server-side fault?
                    example: false
                id:
                    type: string
                    description: ID is a unique identifier for this particular occurrence of the problem.
                    example: 123abc
                message:
                    type: string
                    description: Message is a human-readable explanation specific to this occurrence of the problem.
                    example: parameter 'p' must be an integer
                name:
                    type: string
                    description: Name is the name of this class of errors.
                    example: bad_request
                temporary:
                    type: boolean
                    description: Is the error temporary?
                    example: true
                timeout:
                    type: boolean
                    description: Is the error a timeout?
                    example: true
            description: Bad request
            example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: true
            required:
                - name
                - id
                - message
                - temporary
                - timeout
                - fault
...